Refactor web catalog API

+++ CiRCLES/Views/More/MoreDatabaseAdministratiion.swift
@@ -14,22 +14,23 @@ typealias View = SwiftUI.View
 struct MoreDatabaseAdministratiion: View {
 
     @Environment(AuthManager.self) var authManager
-    @Environment(CatalogManager.self) var catalog
     @Environment(DatabaseManager.self) var database
 
     var body: some View {
         List {
             Section {
                 Button("More.DBAdmin.RedownloadDBs", role: .destructive) {
-                    if let token = authManager.token,
-                       let latestEvent = catalog.latestEvent() {
+                    if let token = authManager.token {
                         withAnimation(.snappy.speed(2.0)) {
                             database.isBusy = true
                         }
                         UIApplication.shared.isIdleTimerDisabled = true
                         Task.detached {
-                            await database.deleteDatabases()
-                            await database.downloadDatabases(for: latestEvent, authToken: token)
+                            if let eventData = await WebCatalog.events(authToken: token),
+                               let latestEvent = eventData.list.first(where: {$0.id == eventData.latestEventID}) {
+                                await database.deleteDatabases()
+                                await database.downloadDatabases(for: latestEvent, authToken: token)
+                            }
                             await MainActor.run {
                                 withAnimation(.snappy.speed(2.0)) {
                                     database.isBusy = false
